+++ NEWS.md
@@ -1,4 +1,4 @@
-# parameters 0.2.1
+# parameters 0.3.0
 
 ## New supported models